+++ CHANGELOG.md
@@ -51,6 +51,7 @@ All notable changes to this project will be documented in this file.
 - *(adapter)* Created a search method for all adapters
 - *(adapter)* Created a questionei adapters for questions
 - *(adapter)* Created a questionei adapter generation context
+- *(analytics)* Created a tracking ai user usage process
 
 ### ⚙️ Miscellaneous Tasks